fix index error
