+++ frontend/app/(auth)/_layout.tsx
@@ -1,9 +1,13 @@
-import { Stack } from 'expo-router';
+import { View, Text, Button } from 'react-native';
+import { useRouter } from 'expo-router';
 
 export default function AuthLayout() {
+  const router = useRouter();
   return (
-    <Stack>
-       <Stack.Screen name="index" options={{ title: 'Home', headerShown: false }} />
-    </Stack>
+    <View style={{ padding: 20 }}>
+          <Text style={{ fontSize: 24 }}>Select Role</Text>
+          <Button title="Admin" onPress={() => router.replace('/(tabs)/(admin)/admin-dashboard')} />
+          <Button title="User" onPress={() => router.replace('/(tabs)/(user)/(tally)/dc')} />
+        </View>
   );
 }
