+++ frontend/app/(tabs)/(admin)/_layout.tsx
@@ -81,7 +81,6 @@ export default function TabLayout() {
           tabBarIcon: ({ color }) => <IconSymbol size={28} name="api" color={color} />,
         }}
       />
-      <Tabs.Screen name="index" options={{ href: null }} />
     </Tabs>
   );
 }
