+++ frontend/app/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from 'expo-router';
+
+export default function AuthLayout() {
+  return (
+    <Stack>
+       <Stack.Screen name="(auth)" options={{ title: 'Auth', headerShown: false }} />
+    </Stack>
+  );
+}
